% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_prototypes.R
\name{new_ExUtil_portfolio}
\alias{new_ExUtil_portfolio}
\title{constructor of EU portfolio object. IEEE 2020}
\usage{
new_ExUtil_portfolio(x, gamma, b)
}
\description{
constructor of EU portfolio object. IEEE 2020
}
\examples{
n<-3e2 # number of realizations
p<-.5*n # number of assets
b<-rep(1/p,p)
gamma<-1

x <- matrix(data = rnorm(n*p), nrow = p, ncol = n)

test <- new_ExUtil_portfolio(x=x, gamma=gamma, b=b)
str(test)
}
