% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{R_GMV}
\alias{R_GMV}
\title{R_GMV. The deterministic value.}
\usage{
R_GMV(mu, Sigma)
}
\arguments{
\item{gamma}{positive numeric. Investors attitude towards risk}
}
\description{
The expected return of the GMV portfolio
}
\examples{
mu <- c(1,5,3,4,9)
sm <- c(1,0,0,0,0,
        0,1,0,0,0,
        0,0,1,0,0,
        0,0,0,1,0,
        0,0,0,0,1)
Sigma <- matrix(data = sm, nrow = p, ncol = p)
}
